+++ tests/testthat/test-families.R
@@ -38,7 +38,7 @@ spde <- make_mesh(loc, c("x", "y"), n_knots = 50, type = "kmeans")
 test_that("Student family fits", {
   skip_on_ci()
   skip_on_cran()
-  set.seed(1)
+  set.seed(3)
   initial_betas <- 0.5
   range <- 0.5
   sigma_O <- 0.3