+++ NEWS.md
@@ -1,8 +1,13 @@
 # sdmTMB
 
-# sdmTMB 0.0.12.9000
+# sdmTMB 0.0.13.9000
 
 * Add IID random intercepts interface. E.g. `... + (1 | g)` #34
+
+# sdmTMB 0.0.12.9000
+
+* Add `epsilon_predictor` argument in `sdmTMB()` to allow a model of the
+  spatiotemporal variance through time.
 
 # sdmTMB 0.0.11.9000